requirements.txt
absl-py==0.9.0
albumentations==0.4.5
autopep8==1.5.4
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
fsspec==0.8.0
future==0.18.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.8.0
imgaug==0.2.6
importlib-metadata==1.7.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
joblib==0.15.1
jupyter-client==6.1.5
jupyter-core==4.6.3
kiwisolver==1.2.0
kornia==0.4.0
Markdown==3.2.2
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
oauthlib==3.1.0
opencv-python-headless==4.2.0.34
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.2.0
six==1.15.0
sklearn==0.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
